7449 : miniassay not deploying correctly on CC machines - don't ensure directory exists before checking if module is modified. if the module is modified, the ensured directory will be deleted anyway
SVN r10580 |2009-03-11 03:05:51 +0000

+++ server/bootstrap/src/org/labkey/bootstrap/ExplodedModule.java
@@ -195,12 +195,15 @@ public class ExplodedModule
     public static void deleteDirectory(File dir)
     {
         //can't delete a directory unless everything inside it is deleted
-        for(File child : dir.listFiles())
+        if (dir.isDirectory())
         {
-            if(child.isDirectory())
-                deleteDirectory(child);
-            else
-                child.delete();
+            for(File child : dir.listFiles())
+            {
+                if(child.isDirectory())
+                    deleteDirectory(child);
+                else
+                    child.delete();
+            }
         }
         
         dir.delete();
